Move inline template model in domain model
