--- src/fastlife/adapters/jinjax/inline.py
@@ -1,5 +1,0 @@
-from fastlife.templates.inline import InlineTemplate
-
-
-class JinjaXTemplate(InlineTemplate):
-    renderer = ".jinja"
