+++ src/fastlife/adapters/jinjax/widgets/base.py
@@ -7,7 +7,7 @@ from typing import Any, Generic, Self, TypeVar
 from markupsafe import Markup
 from pydantic import Field, model_validator
 
-from fastlife.adapters.jinjax.inline import JinjaXTemplate
+from fastlife.domain.model.template import JinjaXTemplate
 from fastlife.services.templates import AbstractTemplateRenderer
 from fastlife.shared_utils.infer import is_union
