+++ src/fastlife/templates/inline.py
@@ -2,30 +2,10 @@
 
 import inspect
 from collections.abc import Callable
-from typing import Any, ClassVar, get_args
+from typing import Any, get_args
 
-from pydantic import BaseModel
-from pydantic.config import ConfigDict
-
+from fastlife.domain.model.template import InlineTemplate
 from fastlife.shared_utils.infer import is_union
-
-
-class InlineTemplate(BaseModel):
-    """
-    Inline templates are used to encourage the location of behavior and the view typing.
-
-    Pages produce templates that are not reusable and don't need to be reusable
-    in there essence, they don't need to be in a component library.
-    They use a component lirary to stay small but contains a view logic
-    tighly coupled with the view and its code can stay in the same module of that view.
-    """
-
-    model_config = ConfigDict(arbitrary_types_allowed=True)
-
-    template: ClassVar[str]
-    """The template string to render."""
-    renderer: ClassVar[str]
-    """Template render engine to use."""
 
 
 def is_inline_template_returned(endpoint: Callable[..., Any]) -> bool:
