+++ tests/fastlife_app/views/app/admin/login.py
@@ -4,7 +4,7 @@ from fastapi.responses import RedirectResponse
 from pydantic import BaseModel, SecretStr
 
 from fastlife import Request, view_config
-from fastlife.adapters.jinjax.inline import JinjaXTemplate
+from fastlife.domain.model.template import JinjaXTemplate
 from fastlife.request.form import FormModel, form_model
 from tests.fastlife_app.config import MyRequest
