+++ tests/fastlife_app/views/app/failed.py
@@ -1,8 +1,8 @@
 from fastapi import Request, Response
 
-from fastlife.adapters.jinjax.inline import JinjaXTemplate
 from fastlife.config.exceptions import exception_handler
 from fastlife.config.views import view_config
+from fastlife.domain.model.template import JinjaXTemplate
 
 
 class MyGoodException(Exception):
