+++ tests/fastlife_app/views/app/home.py
@@ -1,7 +1,7 @@
 from typing import Annotated
 
 from fastlife import view_config
-from fastlife.adapters.jinjax.inline import JinjaXTemplate
+from fastlife.domain.model.template import JinjaXTemplate
 from fastlife.request.form import FormModel, form_model
 from fastlife.request.request import Request
 from tests.fastlife_app.models import Account, Group, Person
