+++ tests/fastlife_app/views/app/i18n/hello_i18n.py
@@ -5,7 +5,7 @@ from typing import Annotated
 from fastapi import Path
 
 from fastlife import view_config
-from fastlife.adapters.jinjax.inline import JinjaXTemplate
+from fastlife.domain.model.template import JinjaXTemplate
 from fastlife.request.localizer import Localizer
